% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distFuncs.R
\name{binPDF}
\alias{binPDF}
\title{Probability mass function for binomial distribution}
\usage{
binPDF(x, mu, var)
}
\arguments{
\item{x}{numeric, quantile}

\item{mu}{numeric, mean}

\item{var}{numeric, variance}
}
\value{
numeric
}
\description{
Binomial PMF, supports fill-rate calculator of \code{\link[=estimateFR]{estimateFR()}}
}
\keyword{internal}
