% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distFuncs.R
\name{detPDF}
\alias{detPDF}
\title{Deterministic probability function}
\usage{
detPDF(x, mu, var)
}
\arguments{
\item{x}{numeric, quantile}

\item{mu}{numeric, mean}

\item{var}{numeric, variance}
}
\value{
numeric
}
\description{
Returns sample mean, supports fill-rate calculator of \code{\link[=estimateFR]{estimateFR()}}
}
\keyword{internal}
