% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventoryPlot.R
\name{estimateFR}
\alias{estimateFR}
\title{Using approach in Silver, et. al, estimate fill rate of CR system}
\usage{
estimateFR(paramDb)
}
\arguments{
\item{paramDb}{data.frame containing Demand, LT, EOQ, ROP}
}
\description{
Using approach in Silver, et. al, estimate fill rate of CR system
}
