% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distFuncs.R
\name{gammaPDF}
\alias{gammaPDF}
\title{Probability mass function for gamma distribution}
\usage{
gammaPDF(x, mu, var)
}
\arguments{
\item{x}{numeric, quantile}

\item{mu}{numeric, mean}

\item{var}{numeric, variance}
}
\value{
numeric
}
\description{
Gamma PDF, supports fill-rate calculator of \code{\link[=esimateFR]{esimateFR()}}
}
\keyword{internal}
