% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventoryPlot.R
\name{getParamDb}
\alias{getParamDb}
\title{Creates a data.frame of parameters suitable for inventory analysis}
\usage{
getParamDb(SL = NULL, EOQ = NULL, Dmd = NULL, LT = NULL, catNames)
}
\arguments{
\item{SL}{numeric, planned minimum safety level}

\item{EOQ}{numeric, Item order quantity}

\item{Dmd}{numeric, Item demand per unit time}

\item{LT}{numeric, lead time}
}
\value{
data.frame
}
\description{
Creates a data.frame of parameters suitable for inventory analysis
}
