% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distFuncs.R
\name{negBinPDF}
\alias{negBinPDF}
\title{Probability mass function for negative binomial distribution}
\usage{
negBinPDF(x, mu, var)
}
\arguments{
\item{x}{numeric, quantile}

\item{mu}{numeric, mean}

\item{var}{numeric, variance}
}
\value{
numeric
}
\description{
Negative binomial PMF, supports fill-rate calculator of \code{\link[=estimateFR]{estimateFR()}}
}
\keyword{internal}
