% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distFuncs.R
\name{normalPDF}
\alias{normalPDF}
\title{Probility density function for normal distribution}
\usage{
normalPDF(x, mu = 1, var = 1)
}
\arguments{
\item{x}{numeric, quantile}

\item{mu}{numeric, mean}

\item{var}{numeric, variance}
}
\value{
numeric
}
\description{
Normal PDF, supports fill-rate calculator of \code{\link[=estimateFR]{estimateFR()}}
}
\keyword{internal}
