% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventoryPlot.R
\name{plotST}
\alias{plotST}
\title{Produces a saw tooth continuous review inventory plot with user supplied parameters}
\usage{
plotST(SL, EOQ = NULL, Dmd = NULL, LT = NULL, cycleLimits = 2,
  catNames = NULL, xMax = NULL, plotLTs = FALSE, FR = TRUE, ...)
}
\arguments{
\item{SL}{numeric, planned minimum safety level to honor or data.frame with at minimum
EOQ, Sl, Dmd, and LT}

\item{EOQ}{Item Order Quantity}

\item{Dmd}{numeric, demand per unit time}

\item{LT}{numeric, lead time}

\item{cycleLimits}{how many saw tooth cycles to draw}

\item{catNames}{names for each sawtooth}

\item{xMax}{if provided provide explicit maximum for x-axis}

\item{plotLTs}{if TRUE provide LT information}

\item{FR}{logical, if TRUE, estimate Fill Rate}

\item{...}{additional parameters, for plot output}
}
\value{
gg
}
\description{
Produces a saw tooth continuous review inventory plot with user supplied parameters
}
\examples{
\dontrun{
plotST(ROP = 4, EOQ = 2, Dmd = 1, LT = 2, plotLTs = TRUE)
plotST(ROP = c(2,3,4), EOQ = c(1,2,3), Dmd = c(2,2,2), LT = 2, plotLTs = TRUE)
plotST(ROP = 4, EOQ = 1.5, Dmd = 1, LT = 2, xMax = 10, plotLTs = FALSE)
plotST(ROP = 4, EOQ = 4, Dmd = 1, LT = 2, cycleLimits = 2, plotLTs = TRUE)

paramDb <- data.frame(ROP = c(2,3,4),
                      EOQ = c(1,2,3),
                      Dmd = c(4,2.5,2.2),
                      LT =  c(1,1,1),
                      Title = paste("Item", 1:3, sep = " "))
plotST(paramDb)

}

}
