% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventoryPlot.R
\name{ppvDecom}
\alias{ppvDecom}
\title{Takes as input PPV parameters.  Returns a data.frame suitable for plotting with \code{\link{plotST}}
Sets ROP to provide a zero safety level cycle
PPV is the procurement problem variable, the average required to meet the demand over the net leadtime
of an item with both a repairable pipeline (regenerative) and a procurement pipeline (attrition)}
\usage{
ppvDecom(SL = 0, Demand, CRR, SR, PCLT, PRTAT, Cov_Dur = 2,
  Repair_period = 1)
}
\arguments{
\item{SL}{numeric, minimum stock level}

\item{Demand}{numeric, quarterly demand forecast}

\item{CRR}{numeric, proportion of items that are returned for repair}

\item{SR}{numeric, proportion of items that are successfully repaired}

\item{PCLT}{numeric, production lead time in quarters}

\item{PRTAT}{numeric, process repair turn around time in quarters}

\item{Cov_Dur}{numeric, number of quarters of material to buy}

\item{Repair_period}{number, number of quarters of material to repair}
}
\value{
data.frame
}
\description{
Takes as input PPV parameters.  Returns a data.frame suitable for plotting with \code{\link{plotST}}
Sets ROP to provide a zero safety level cycle
PPV is the procurement problem variable, the average required to meet the demand over the net leadtime
of an item with both a repairable pipeline (regenerative) and a procurement pipeline (attrition)
}
