% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventoryPlot.R
\name{ppvDecom}
\alias{ppvDecom}
\title{Takes as input PPV parmeters.  Returns a data.frame suitable for plotting with \link{\code{plotST}}}
\usage{
ppvDecom(Dmd, SRR, CR, PCLT, PRTAT)
}
\arguments{
\item{Dmd}{numeric, quarterly demand forecast}

\item{CR}{numeric, proportion of items that are returned for repair}

\item{PCLT}{numeric, production lead time in quarters}

\item{PRTAT}{numeric, process repair turn around time in quarters}

\item{SR}{numeric, proportion of items that are successful repaired}
}
\value{
data.frame
}
\description{
PPV is the procurement problem variable, the average required to meet the demand over the net leadtime
of an item with both a repairable pipeline (regenerative) and a procurement pipeline (attrition)
}
